+++ appdaemon/adapi.py
@@ -5,6 +5,7 @@ import iso8601
 import re
 from datetime import timedelta
 from copy import deepcopy
+from typing import Any, Optional, Callable
 
 # needed for fake coro cb that looks like scheduler
 import uuid
@@ -1534,11 +1535,13 @@ class ADAPI:
     #
 
     @staticmethod
-    def _check_service(service):
+    def _check_service(service: str) -> None:
         if service.find("/") == -1:
             raise ValueError("Invalid Service Name: {}".format(service))
 
-    def register_service(self, service, cb, **kwargs):
+    def register_service(
+        self, service: str, cb: Callable[[str, str, str, dict], Any], **kwargs: Optional[dict]
+    ) -> None:
         """Registers a service that can be called from other apps, the REST API and the Event Stream
 
         Using this function, an App can register a function to be available in the service registry.
@@ -1550,6 +1553,10 @@ class ADAPI:
             cb: A reference to the function to be called when the service is requested. This function may be a regular
                 function, or it may be async. Note that if it is an async function, it will run on AppDaemon's main loop
                 meaning that any issues with the service could result in a delay of AppDaemon's core functions.
+        Keyword Args:
+            namespace(str, optional): Namespace to use for the call. See the section on
+                `namespaces <APPGUIDE.html#namespaces>`__ for a detailed description.
+                In most cases, it is safe to ignore this parameter.
 
         Returns:
             None
@@ -1571,7 +1578,42 @@ class ADAPI:
 
         self.AD.services.register_service(namespace, d, s, cb, __async="auto", **kwargs)
 
-    def list_services(self, **kwargs):
+    def deregister_service(self, service: str, **kwargs: Optional[dict]) -> bool:
+        """Deregisters a service that had been previously registered
+
+        Using this function, an App can deregister a service call, it has initially registered in the service registry.
+        This will automatically make it unavailable to other apps using the `call_service()` API call, as well as published
+        as a service in the REST API and make it unavailable to the `call_service` command in the event stream.
+        This function can only be used, within the app that registered it in the first place
+
+        Args:
+            service: Name of the service, in the format `domain/service`.
+        Keyword Args:
+            namespace(str, optional): Namespace to use for the call. See the section on
+                `namespaces <APPGUIDE.html#namespaces>`__ for a detailed description.
+                In most cases, it is safe to ignore this parameter.
+
+        Returns:
+            Bool
+
+        Examples:
+            >>> self.deregister_service("myservices/service1")
+
+        """
+        self._check_service(service)
+        d, s = service.split("/")
+        self.logger.debug("deregister_service: %s/%s, %s", d, s, kwargs)
+
+        namespace = self._get_namespace(**kwargs)
+
+        if "namespace" in kwargs:
+            del kwargs["namespace"]
+
+        kwargs["__name"] = self.name
+
+        return self.AD.services.deregister_service(namespace, d, s, **kwargs)
+
+    def list_services(self, **kwargs: Optional[dict]) -> list:
         """List all services available within AD
 
         Using this function, an App can request all available services within AD
@@ -1604,7 +1646,7 @@ class ADAPI:
         return self.AD.services.list_services(namespace)  # retrieve services
 
     @utils.sync_wrapper
-    async def call_service(self, service, **kwargs):
+    async def call_service(self, service: str, **kwargs: Optional[dict]) -> Any:
         """Calls a Service within AppDaemon.
 
         This function can call any service and provide any required parameters.