+++ docs/HISTORY.rst
@@ -7,6 +7,8 @@ Change Log
 **Features**
 
 - Added "get_logs" command to stream
+- Added "deregister_service" api call
+- Added new AD event `service_deregistered`. This is fired when an app's service is deregistered
 - Added the use of the ``!include`` directive in AD. This can be used either in the main ``appdaemon`` or ``apps`` config
 - Added support for Python 3.9
 - Allowed for service calls over REST API to return data
@@ -23,6 +25,7 @@ Change Log
 - Fixed issue with when a plugin that is persistent re-initializes, and it creates an error
 - Fixed issue with when an entity has no state, and if wanting to listen to it, breaks internally
 - Fixed a couple of scheduler issues that affected tmezones west of EDT
+- Ensured that when apps with registered services are terminated, their services are also deregistered
 - Documentation fixes - contributed by `sithmein <https://github.com/sithmein>`__
 - Fixed an issue where AD crashes when multiple plugins restart at the same time